requirements.txt
pandas
numpy
matplotlib
seaborn
statsmodels.api